Create LexicalAnalyser.c
The lexical analyzer is the part of the compiler that detects the token of the program and sends it to the syntax analyzer. It converts the input program into a sequence of Tokens. 
A C program consists of various tokens and a token is either a keyword, an identifier, a constant, a string literal, or a symbol.

For Example: 

1) Keywords: 
Examples- for, while, if etc.

2) Identifier
Examples- Variable name, function name etc.

3) Operators:
Examples- '+', '++', '-' etc.

4) Separators:
Examples- ', ' ';' etc

It is a C program to print all the keywords, literals, valid identifiers, invalid identifiers, integer number, real number in a given C program.




















Examples of different types of tokens in C.

Example
Keywords: for, if, include, etc
Identifier: variables, functions, etc
separators: ‘,’, ‘;’, etc
operators: ‘-’, ‘=’, ‘++’, etc